Create rql-data-model tests

+++ examples/node/v12/__tests__/rql-data-models.test.js
@@ -1,54 +1,63 @@
-import Realm from "realm";
-import { ItemModel, ProjectModel } from "./schemas/rql-data-models";
+import Realm, { BSON } from "realm";
+import { Item, Project } from "./models/rql-data-models.ts";
 
-describe("test models", () => {
+describe("Test Item and Project RQL Models", () => {
   let realm;
-  const config = {
-    schema: [ItemModel, ProjectModel],
-    path: "testing.realm",
-  };
+  const path = "testing.realm";
+
   beforeEach(async () => {
-    realm = await Realm.open(config);
-  });
-  afterEach(() => {
-    realm.write(() => {
-      realm.deleteAll();
+    realm = await Realm.open({
+      schema: [Project, Item],
+      inMemory: true,
+      path: path,
     });
-    realm.close();
-    expect(realm.isClosed).toBe(true);
   });
+
+  afterEach(() => {
+    // After each test, delete the objects and close the realm
+    if (realm && !realm.isClosed) {
+      realm.write(() => {
+        realm.deleteAll();
+      });
+      realm.close();
+      expect(realm.isClosed).toBe(true);
+    }
+  });
+
   afterAll(() => {
-    Realm.deleteFile(config);
+    Realm.deleteFile({ path });
   });
-  test("open realm with config", async () => {
+  test("Can open realm with config", async () => {
     expect(realm.isClosed).toBe(false);
   });
+
   test("Can create object of Item type", () => {
     realm.write(() => {
-      realm.create("Item", {
-        id: new Realm.BSON.ObjectId(),
+      realm.create(Item, {
+        _id: new BSON.ObjectId(),
         name: "get coffee",
       });
     });
-    const coffeeItem = realm.objects("Item")[0];
-    expect(coffeeItem.id instanceof Realm.BSON.ObjectId).toBe(true);
+    const coffeeItem = realm.objects(Item)[0];
+    expect(coffeeItem._id instanceof BSON.ObjectId).toBe(true);
     expect(coffeeItem.name).toBe("get coffee");
     expect(coffeeItem.isComplete).toBe(false);
   });
+
   test("Can create object of Project type", () => {
     realm.write(() => {
-      const teaItem = realm.create("Item", {
-        id: new Realm.BSON.ObjectId(),
+      const teaItem = realm.create(Item, {
+        _id: new BSON.ObjectId(),
         name: "get tea",
       });
-      realm.create("Project", {
-        id: new Realm.BSON.ObjectId(),
+      realm.create(Project, {
+        _id: new BSON.ObjectId(),
         name: "beverages",
         items: [teaItem],
       });
     });
-    const bevProject = realm.objects("Project")[0];
-    expect(bevProject.id instanceof Realm.BSON.ObjectId).toBe(true);
+    const bevProject = realm.objects(Project)[0];
+    expect(bevProject._id instanceof BSON.ObjectId).toBe(true);
     expect(bevProject.name).toBe("beverages");
     expect(bevProject.items[0].name).toBe("get tea");
   });
